backend_hackaton-main/BACKEND/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings>=2.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
earthengine-api==0.1.375
rasterio>=1.3.9
xarray>=2023.12.0
numpy>=1.26.0
pandas>=2.0.3
scikit-learn>=1.3.2
opencv-python>=4.8.1.78
pillow>=10.1.0
matplotlib>=3.8.2
seaborn>=0.13.0
folium>=0.15.0
plotly>=5.17.0
python-dotenv>=1.0.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
setuptools>=65.0.0
tensorflow>=2.15.0
joblib>=1.3.2
